states50_labels shapefile data

Point coordinates for the center of the hexagon that includes: 50 States only

Usage

states50_labels

Format

Simple feature collection

abbr_usps

2-letter code used by the United States Postal Service

abbr_gpo

abbreviations from US Government Printing Office

abbr_ap

abbreviations from AP Stylebook

name

Name of geographic area

fips

2-digit FIPS code (Federal Information Processing Standard)

geo_type

Type of region: state, federal district, territory, freely associated state

geometry

point coordinates for the center of the hexagon

Source

Coordinates for hex map are based on NPR's daily graphic template, https://github.com/nprapps/dailygraphics-templates/blob/master/state_grid_map/index.html

Abbreviations are pulled from List of U.S. state and territory abbreviations, https://en.wikipedia.org/wiki/List_of_U.S._state_and_territory_abbreviations