usa52_labels shapefile data

Point coordinates for the center of the hexagon that includes:

50 States

1 federal district: District of Columbia

1 territory: Puerto Rico

Usage

usa52_labels

Format

Simple feature collection

abbr_usps

2-letter code used by the United States Postal Service

abbr_gpo

abbreviations from US Government Printing Office

abbr_ap

abbreviations from AP Stylebook

abbr_short

abbreviations that are between 3-4 characters

abbr_long

abbreviations that are between 4-8 characters, most are 5-6 characters

name

Name of geographic area

fips

2-digit FIPS code (Federal Information Processing Standard)

geo_type

Type of region: state, federal district, territory, freely associated state

geometry

point coordinates for the center of the hexagon